Fix data type that we submit to the data stream.

+++ catkit2/services/thorlabs_pm_sim/thorlabs_pm_sim.py
@@ -13,7 +13,7 @@ class ThorlabsPMSim(Service):
     def main(self):
         while not self.should_shut_down:
             power = self.get_power()
-            self.power.submit_data(np.array([power]))
+            self.power.submit_data(np.array([power], dtype='float64'))
 
             self.sleep(self.interval)
